+++ collections/hashmap/contracts/UnsafeHashmap.sol
@@ -41,4 +41,17 @@ contract UnsafeHashmap is Util {
   function size() public constant returns (uint) {
     return values.length -1; // not counting entry 0
   }
+
+  function remove(string _key) public {
+    uint index = keyMap[b32(_key)];
+    if (index == 0) return;
+    // remove the index mapping
+    keyMap[b32(_key)] = 0;
+    // remove the value
+    values[index] = 0;
+    // remove the key
+    if (isIterable) {
+      delete keys[index];
+    }
+  }
 }